Set color_scheme variable in default.html
If default_color_scheme variable is defined and set to light, color
scheme will be set to light.

If default_color_scheme variable is not defined or set to black or set
to an unknown value, color scheme will be set to dark.

color_scheme variable should be set to either "theme--light" or
"theme--dark" (the only two valid class names for themes).