Hard-code CSS headings font-sizes
Lighthouse says that h3-h6 elements do not have a sufficient contrast
ratio. After trying multiple values manually, I found that 1.5em, which
is equivalent to 24px for this particular style-sheet, is the smallest
acceptable value by Lighthouse.

To fix this, I changed all headings font-sizes to be between 1.5em and
2em (2em is the default value for h1 in Chrome) with a step of 0.1em.

TODO: This solution is not elegant.