Add a new post_description_animation variable in _config.yml
If this variable is set to false then post description animation will be
disabled. The default behavior (when this variable is not set, or set to
anything else different from false) is to animate the post description.

This variable is useful for those who want to improve their Lighthouse
score (on performance). In fact, my Lighthouse score is 97% when post
description is animated, and is 99% or 100% when it is disabled.